[+] Added test for removal random and weight random

+++ Assets/JayTools/JayRandoms/Instance/WeightedItem.cs
@@ -8,14 +8,14 @@
     public class WeightedItem <T>
     {
         public T Item;
-        public int Weight;
+        public float Weight;
 
         /// <summary>
         /// A constructor of Weighted Item.
         /// </summary>
         /// <param name="item"></param>
         /// <param name="weight"></param>
-        public WeightedItem(T item, int weight)
+        public WeightedItem(T item, float weight)
         {
             Item = item;
             Weight = weight;
